\name{active_regime}
\alias{active_regime}
\title{Return the value of regime or NA}
\arguments{
  \item{regime}{The current regime}

  \item{.lambda.r_1}{Boolean to control switching}
}
\description{
  Return the value of regime or NA
}
\examples{
active_regime(r, TRUE)
}
\author{
  Brian Lee Yung Rowe
}
\keyword{manip}

